Re-architected api as requested.
