+++ backend/src/admin/admin.controller.ts
@@ -2,32 +2,26 @@ import { Controller, Post, Body, UsePipes, ValidationPipe, UseGuards } from '@ne
 import { AuthGuard } from '@nestjs/passport';
 import { RegisterAdminDto, LoginAdminDto } from './dto';
 import { AdminService } from './admin.service';
-import { RegisterJuniorDto } from '../junior/dto';
 import { Admin } from './admin.decorator';
+import { AuthenticationService } from '../authentication/authentication.service';
 
 @Controller('admin')
 export class AdminController {
   constructor(
     private readonly adminService: AdminService,
+    private readonly authenticationService: AuthenticationService,
   ) { }
-
-  @UsePipes(new ValidationPipe())
-  @Post('register/admin')
-  async create(@Body() userData: RegisterAdminDto) {
-    return this.adminService.register(userData);
-  }
 
   @UsePipes(new ValidationPipe())
   @Post('login')
   async login(@Body() userData: LoginAdminDto) {
-    return this.adminService.login(userData);
+    return this.authenticationService.loginAdmin(userData);
   }
 
   @UsePipes(new ValidationPipe())
-  @UseGuards(AuthGuard('jwt'))
-  @Post('register/junior')
-  async registerJunior(@Admin() payload: any, @Body() userData: RegisterJuniorDto) {
-    await this.adminService.verifyIsAdmin(payload.user);
-    return await this.adminService.registerJunior(userData);
+  @Post('register')
+  async create(@Body() userData: RegisterAdminDto) {
+    return this.adminService.registerAdmin(userData);
   }
+
 }
