+++ backend/src/junior/junior.service.ts
@@ -1,9 +1,11 @@
-import { Injectable, Inject, forwardRef } from '@nestjs/common';
+import { Injectable, Inject, forwardRef, ConflictException } from '@nestjs/common';
 import { Junior } from './junior.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
-import { LoginJuniorDto } from './dto';
-import { AuthenticationService } from '../authentication/authentication.service';
+import { RegisterJuniorDto } from './dto';
+import { hash } from 'bcrypt';
+import { saltRounds } from '../authentication/authentication.consts';
+import * as content from '../content.json';
 
 @Injectable()
 export class JuniorService {
@@ -11,11 +13,7 @@ export class JuniorService {
     constructor(
         @InjectRepository(Junior)
         private readonly juniorRepo: Repository<Junior>,
-        @Inject(forwardRef(() => AuthenticationService))
-        private readonly authenticationService: AuthenticationService,
     ) { }
-
-    login = async (loginData: LoginJuniorDto) => this.authenticationService.loginJunior(loginData);
 
     async getUser(phoneNumber: string): Promise<Junior> {
         return await this.juniorRepo.findOne({ phoneNumber });
@@ -23,6 +21,24 @@ export class JuniorService {
 
     async createUser(details: Junior) {
         await this.juniorRepo.save(details);
+    }
+
+    /**
+  Currently this returns the pin as we need pass that back to frontend.
+  Will be corrected when relevant workflow is introduced.
+ */
+    async registerJunior(registrationData: RegisterJuniorDto): Promise<string> {
+        const userExists = await this.getUser(registrationData.phoneNumber);
+        if (userExists) { throw new ConflictException(content.AdminAlreadyExists); }
+        const pin = this.generatePin();
+        const hashedPassword = await hash(pin, saltRounds);
+        const junior = {
+            firstName: registrationData.firstName, lastName: registrationData.lastName,
+            phoneNumber: registrationData.phoneNumber, pin: hashedPassword,
+        } as Junior;
+        await this.createUser(junior);
+        // return `${registrationData.phoneNumber} ${content.Created} (PIN:${pin})`;
+        return pin.toString();
     }
 
     // This will be moved to its own service when the pin workflow is produced.
